third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/module/dynamic-import/dynamic-imports-credentials.sub-expected.txt

This is a testharness.js-based test.
FAIL Dynamic imports should be loaded with or without the credentials based on the same-origin-ness and the parent script's crossOrigin attribute assert_equals: Dynamic descendant modules should be loaded with the credentials when the crossOrigin attribute is not specified and the target is same-origin expected "found" but got "not found"
Harness: the test ran to completion.
